factor out application resolver
